+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ All notable changes to this project are documented in this file.
 - Update mainnet bootstrap files
 - Fix size calculations for all serializable classes
 - Add ``size`` key to JSON output of Block and Transaction
+- add prompt command to split VIN to multiple VOUT
 
 [0.7.3] 2018-07-12
 ------------------